- Fixed some function definitions so it will be easier to use 64-bit integers in the simulation/function runtime
git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@7152 f25d12d1-65f4-0310-ae8a-bbce733d8d8e

+++ c_runtime/meta_modelica_builtin.cpp
@@ -191,7 +191,7 @@ intString_rettype intString(modelica_integer i)
 {
   /* 64-bit integer: 1+log_10(2**63)+1 = 20 digits max */
   static char buffer[32];
-  sprintf(buffer, "%ld", i);
+  sprintf(buffer, "%ld", (long) i);
   return strdup(buffer);
 }
